update folder name
