+++ src/vg_ecg_detectors/vg_detector.py
