--- src/vg_ecg_detectors_emrich/__init__.py
@@ -1,3 +1,0 @@
-from vg_ecg_detectors_emrich.vg_detector import VisGraphDetector
-from vg_ecg_detectors_emrich.fast_whvg import FastWHVG
-from vg_ecg_detectors_emrich.fast_nvg import FastNVG
